Use java nio for multiple clients & smaller audio chunks
This allows more than one parent to listen in with their device.
The overhead should be minimal, only the traffic should increase
by the number of devices.

Additionally, the main event-loop is now rather being driven by reading
the audio stream, which will read only audio-data for a fixed
time-interval, and write that to all sockets where data can be written
to.

If no socket can receive the data, then instead of blocking and
accumulating data in the audio-buffer, we simply drop this chunk.

This ensures, that the clients will receive the newest audio data
we hold.

The client buffer size will also be reduced to make use of that
behaviour on the client side. Not that it is likely to happen, as the
buffers apparently cannot be set *that* low.